Corectly handle frontend attribute mapping to keep editor working
