+++ assets/js/atomic/blocks/product/button/block.js
@@ -24,7 +24,7 @@ import './style.scss';
  *                                   this is not provided.
  * @return {*} The component.
  */
-const ProductButton = ( { className, ...props } ) => {
+const Block = ( { className, ...props } ) => {
 	const productDataContext = useProductDataContext();
 	const product = props.product || productDataContext.product;
 
@@ -133,9 +133,9 @@ const AddToCartButtonPlaceholder = () => {
 	);
 };
 
-ProductButton.propTypes = {
+Block.propTypes = {
 	className: PropTypes.string,
 	product: PropTypes.object,
 };
 
-export default ProductButton;
+export default Block;
