+++ assets/js/base/hocs/with-filtered-attributes.js
@@ -1,39 +1,7 @@
 /**
- * Given some block attributes, gets attributes from the dataset or uses defaults.
- *
- * @todo Importing this from base-utils causes errors in the editor - find out why.
- *
- * @param {Object} blockAttributes Object containing block attributes.
- * @param {Array} rawAttributes Dataset from DOM.
- * @return {Array} Array of parsed attributes.
+ * External dependencies
  */
-const getValidBlockAttributes = ( blockAttributes, rawAttributes ) => {
-	const attributes = [];
-
-	Object.keys( blockAttributes ).forEach( ( key ) => {
-		if ( typeof rawAttributes[ key ] !== 'undefined' ) {
-			switch ( blockAttributes[ key ].type ) {
-				case 'boolean':
-					attributes[ key ] = rawAttributes[ key ] !== 'false';
-					break;
-				case 'number':
-					attributes[ key ] = parseInt( rawAttributes[ key ], 10 );
-					break;
-				case 'array':
-				case 'object':
-					attributes[ key ] = JSON.parse( rawAttributes[ key ] );
-					break;
-				default:
-					attributes[ key ] = rawAttributes[ key ];
-					break;
-			}
-		} else {
-			attributes[ key ] = blockAttributes[ key ].default;
-		}
-	} );
-
-	return attributes;
-};
+import { getValidBlockAttributes } from '@woocommerce/base-utils';
 
 /**
  * HOC that filters given attributes by valid block attribute values, or uses defaults if undefined.
