+++ assets/js/base/utils/index.js
@@ -4,3 +4,4 @@ export * from './address';
 export * from './shipping-rates';
 export * from './legacy-events';
 export * from './render-frontend';
+export * from './get-valid-block-attributes';
