+++ assets/js/base/utils/render-frontend.js
@@ -5,41 +5,6 @@ import { render } from 'react-dom';
 import BlockErrorBoundary from '@woocommerce/base-components/block-error-boundary';
 
 /**
- * Given some block attributes, gets attributes from the dataset or uses defaults.
- *
- * @param {Object} blockAttributes Object containing block attributes.
- * @param {Array} rawAttributes Dataset from DOM.
- * @return {Array} Array of parsed attributes.
- */
-export const getValidBlockAttributes = ( blockAttributes, rawAttributes ) => {
-	const attributes = [];
-
-	Object.keys( blockAttributes ).forEach( ( key ) => {
-		if ( typeof rawAttributes[ key ] !== 'undefined' ) {
-			switch ( blockAttributes[ key ].type ) {
-				case 'boolean':
-					attributes[ key ] = rawAttributes[ key ] !== 'false';
-					break;
-				case 'number':
-					attributes[ key ] = parseInt( rawAttributes[ key ], 10 );
-					break;
-				case 'array':
-				case 'object':
-					attributes[ key ] = JSON.parse( rawAttributes[ key ] );
-					break;
-				default:
-					attributes[ key ] = rawAttributes[ key ];
-					break;
-			}
-		} else {
-			attributes[ key ] = blockAttributes[ key ].default;
-		}
-	} );
-
-	return attributes;
-};
-
-/**
  * Renders a block component in the place of a specified set of selectors.
  *
  * @param {Object}   props                         Render props.
